requirements.txt
Flask
flask-cors
yt-dlp
webvtt-py
sumy
requests
gunicorn
